requirements.txt
bokeh

selenium
pillow
pyyaml
